refactor(selection): move outlines and resize handles into aboveView (Phase B / B′)
Selection outlines and resize handles relocated from bgView to
aboveView. Phase B′ corrects the body-layer hit walk to be
front-to-back per entity z-order.
