+++ tests/unit/hit-test.test.ts
@@ -168,3 +168,68 @@ describe('hit-test — chrome only on chrome-bearing entities', () => {
     expect(result.payload).toEqual({ kind: 'background' })
   })
 })
+
+describe('hit-test — body z-order (front-to-back)', () => {
+  // entityOrder semantics: array order is back-to-front (paint order — the
+  // last item paints on top, matching JSON Canvas v1.0). The hit-test must
+  // walk the body layer front-to-back so a sticky painted over a frame
+  // resolves to the sticky, not the frame body underneath.
+  //
+  // Frame body at (200,200), 400×300. Sticky at (300,300), 100×40.
+  // The sticky is fully inside the frame.
+
+  it('sticky declared front (after frame in entities) wins over frame body', () => {
+    const f = frame({ id: 'f1', screenX: 200, screenY: 200 })
+    const t = text('t1', 300, 300)
+    // Frame first (back), text after (front).
+    const result = hitTest(inputs([f, t]), { x: 320, y: 320 })
+    expect(result.layer).toBe('body')
+    expect(result.payload).toEqual({
+      kind: 'entity-body',
+      entityId: 't1',
+      entityKind: 'text',
+    })
+  })
+
+  it('reverse z-order (sticky behind frame) returns frame-body', () => {
+    const f = frame({ id: 'f1', screenX: 200, screenY: 200 })
+    const t = text('t1', 300, 300)
+    // Text first (back), frame after (front).
+    const result = hitTest(inputs([t, f]), { x: 320, y: 320 })
+    expect(result.layer).toBe('body')
+    expect(result.payload).toEqual({ kind: 'frame-body', entityId: 'f1' })
+  })
+
+  it('two non-group entities — last in entities wins (front)', () => {
+    // Two stacked text entities both covering the click point.
+    const t1 = text('t1', 200, 200, 200, 200)
+    const t2 = text('t2', 200, 200, 200, 200)
+    // t2 declared after t1 → front.
+    const result = hitTest(inputs([t1, t2]), { x: 300, y: 300 })
+    expect(result.payload).toEqual({
+      kind: 'entity-body',
+      entityId: 't2',
+      entityKind: 'text',
+    })
+    // Reversed declaration order swaps which wins.
+    const reversed = hitTest(inputs([t2, t1]), { x: 300, y: 300 })
+    expect(reversed.payload).toEqual({
+      kind: 'entity-body',
+      entityId: 't1',
+      entityKind: 'text',
+    })
+  })
+
+  it('group containment still wins regardless of declared order', () => {
+    // Even if the group is declared LAST (front-most by paint order), a
+    // member entity inside the group must hit first — groups are containers,
+    // their hit-test runs after non-group bodies.
+    const g = group('g1', 100, 100, 600, 500)
+    const t = text('t1', 300, 300)
+    // Group declared after text → would be "front" by paint order, but
+    // groups sit at the bottom of the hit-test priority within the body
+    // layer.
+    const result = hitTest(inputs([t, g]), { x: 350, y: 320 })
+    expect(result.payload).toMatchObject({ kind: 'entity-body', entityId: 't1' })
+  })
+})
